# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DeenMate API
  description: |
    Comprehensive Islamic application backend providing access to Quran, Hadith, Prayer Times, 
    Zakat calculations, and Islamic audio content. This API is designed to serve Islamic 
    applications, websites, and mobile apps with accurate, reliable, and well-structured Islamic data.
  
    ## Features
    - **Quran Module**: Complete Quran text, translations, and recitations
    - **Prayer Module**: Accurate prayer times, qibla direction, and calculation methods
    - **Hadith Module**: Authentic hadith collections and search functionality
    - **Zakat Module**: Zakat calculations and current market values
    - **Audio Module**: High-quality Islamic audio content and metadata
  
    ## Authentication
    Currently operates in read-only mode without authentication. Future versions will include 
    JWT-based authentication for user-specific features.
  
    ## Rate Limiting
    - Public Endpoints: 100 requests per minute per IP
    - Authenticated Endpoints: 1000 requests per minute per user
    - Premium Tier: 5000 requests per minute per user
  
    ## Support
    - Documentation: https://docs.deenmate.com
    - API Status: https://status.deenmate.com
    - Support Email: api-support@deenmate.com
    - Developer Discord: https://discord.gg/deenmate
  
  version: 1.0.0
  contact:
    name: DeenMate API Support
    email: api-support@deenmate.com
    url: https://docs.deenmate.com
  license:
...
servers:
...
paths:
...
  /hadith/collections/{id}/books:
    get:
      tags:
        - Hadith
      summary: Get Books in Collection
      description: Retrieve books from a specific hadith collection
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: integer
      responses:
        '200':
          description: List of books in the collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithBooksResponse'
  /hadith/collections/{collectionId}/books/{bookId}/hadiths:
    get:
      tags:
        - Hadith
      summary: Get Hadiths in Book
      description: Retrieve hadiths from a specific book with pagination
      parameters:
        - name: collectionId
          in: path
          required: true
          description: Collection ID
          schema:
            type: integer
        - name: bookId
          in: path
          required: true
          description: Book ID
          schema:
            type: integer
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Hadiths per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Hadiths with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithsResponse'
  /hadith/search:
    get:
      tags:
        - Hadith
      summary: Search Hadiths
      description: Search hadiths across collections
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: collection
          in: query
          description: Collection ID to search within
          schema:
            type: integer
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Results per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithSearchResponse'
        '400':
          description: Missing search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /hadith/{id}:
    get:
      tags:
        - Hadith
      summary: Get Hadith by ID
      description: Retrieve a specific hadith by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Hadith ID
          schema:
            type: integer
      responses:
        '200':
          description: Hadith details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithResponse'
        '404':
          description: Hadith not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /audio/verse/{reciterId}/{chapterId}/{verseNumber}:
    get:
      tags:
      - Quran
      summary: Get Verse Audio
      description: Retrieve audio metadata for a specific verse
      parameters:
        - name: reciterId
          in: path
          required: true
          description: Reciter ID
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          description: Chapter ID
          schema:
            type: integer
        - name: verseNumber
          in: path
          required: true
          description: Verse number
          schema:
            type: integer
        - name: quality
          in: query
          description: Audio quality
          schema:
            type: string
            enum: [64kbps, 128kbps, 192kbps, 320kbps]
            default: 128kbps
      responses:
        '200':
          description: Verse audio metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerseAudioResponse'
  /audio/chapter/{reciterId}/{chapterId}:
    get:
      tags:
      - Quran
      summary: Get Chapter Audio
      description: Retrieve audio metadata for a complete chapter
      parameters:
        - name: reciterId
          in: path
          required: true
          description: Reciter ID
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          description: Chapter ID
          schema:
            type: integer
        - name: quality
          in: query
          description: Audio quality
          schema:
            type: string
            enum: [64kbps, 128kbps, 192kbps, 320kbps]
            default: 128kbps
      responses:
        '200':
          description: Chapter audio metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterAudioResponse'
...
  /audio/reciter/{reciterId}/stats:
    get:
      tags:
      - Quran
      summary: Get Reciter Stats
      description: Retrieve audio statistics for a specific reciter
      parameters:
        - name: reciterId
          in: path
          required: true
          description: Reciter ID
          schema:
            type: integer
      responses:
        '200':
          description: Reciter audio statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReciterStatsResponse'
  /audio/search:
    get:
      tags:
      - Quran
      summary: Search Audio
      description: Search audio content
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: reciter
          in: query
          description: Reciter ID
          schema:
            type: string
        - name: quality
          in: query
          description: Audio quality
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: string
            default: "1"
        - name: per_page
          in: query
          description: Results per page
          schema:
            type: string
            default: "20"
      responses:
        '200':
          description: Search results with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioSearchResponse'
        '400':
          description: Missing search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audio/validate:
    get:
      tags:
      - Quran
      summary: Validate Audio URL
      description: Validate if an audio URL is still valid
      parameters:
        - name: url
          in: query
          required: true
          description: Audio URL to validate
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: URL validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlValidationResponse'
        '400':
          description: Missing URL parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          description: Response data
        meta:
          description: Metadata about the response
      required:
        - success
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: false
            error:
              type: string
              description: Error message
            code:
              type: string
              description: Error code
            details:
              description: Additional error details
          required:
            - error
...
    AudioMetadata:
      type: object
      properties:
        id:
          type: integer
          description: Unique audio ID
        reciterId:
          type: integer
          description: Reciter ID
        chapterId:
          type: integer
          description: Chapter ID
        verseNumber:
          type: integer
          description: Verse number (for verse audio)
        quality:
          type: string
          description: Audio quality
        format:
          type: string
          description: Audio format
        duration:
          type: integer
          description: Duration in seconds
        fileSize:
          type: integer
          description: File size in bytes
        url:
          type: string
          format: uri
          description: Audio file URL
        expiresAt:
          type: string
          format: date-time
          description: URL expiration time
    VerseAudioResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: true
            data:
              type: object
              properties:
                audioMetadata:
                  $ref: '#/components/schemas/AudioMetadata'
            meta:
              type: object
              properties:
                type:
                  type: string
                  const: verse
                reciterId:
                  type: integer
                chapterId:
                  type: integer
                verseNumber:
                  type: integer
                quality:
                  type: string
                cacheTtl:
                  type: string

  # Security Schemes (for future use)
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated endpoints

# Global security (for future use)
# security:
#   - BearerAuth: []